refactor authService
