+++ lib/app/modules/splash_screen/splash_screen_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:habitat_ft_user/app/modules/login/services/auth_service.dart';
+import 'package:habitat_ft_user/app/modules/login/login_controller.dart';
 
 import 'package:habitat_ft_user/app/modules/splash_screen/splash_screen_controller.dart';
 
@@ -7,6 +7,6 @@ class SplashScreenBinding extends Bindings {
   @override
   void dependencies() {
     Get.put<SplashScreenController>(SplashScreenController());
-    Get.put<AuthService>(AuthService());
+    Get.put<LoginController>(LoginController());
   }
 }
