+++ lib/app/modules/splash_screen/splash_screen_controller.dart
@@ -1,13 +1,13 @@
 import 'dart:async';
 
 import 'package:get/get.dart';
-import 'package:habitat_ft_user/app/modules/login/services/auth_service.dart';
+import 'package:habitat_ft_user/app/modules/login/login_controller.dart';
 
 class SplashScreenController extends GetxController {
   @override
   void onInit() {
     Timer(Duration(seconds: 3), () {
-      Get.find<AuthService>().fetch();
+      Get.find<LoginController>().fetch();
     });
   }
