+++ lib/app/modules/splash_screen/splash_screen_view.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:habitat_ft_user/app/modules/login/services/auth_service.dart';
+import 'package:habitat_ft_user/app/modules/login/login_controller.dart';
 import 'package:habitat_ft_user/app/modules/splash_screen/splash_screen_controller.dart';
 import 'package:habitat_ft_user/app/utils/config/assets.dart';
 
 class SplashScreenView extends GetView<SplashScreenController> {
   @override
   Widget build(BuildContext context) {
-    Get.find<AuthService>();
+    Get.find<LoginController>();
     return Stack(
       children: [
         Container(
